Add resource delete/add notifications + cleanup

+++ src/Controller/Resources/ResourcesDeleteController.php
@@ -17,10 +17,14 @@ namespace App\Controller\Resources;
 
 use App\Controller\AppController;
 use App\Error\Exception\ValidationRuleException;
+use App\Model\Entity\Role;
 use Cake\Datasource\Exception\RecordNotFoundException;
+use Cake\Event\Event;
 use Cake\Network\Exception\BadRequestException;
 use Cake\Network\Exception\ForbiddenException;
+use Cake\Network\Exception\InternalErrorException;
 use Cake\Network\Exception\NotFoundException;
+use Cake\ORM\TableRegistry;
 use Cake\Validation\Validation;
 
 class ResourcesDeleteController extends AppController
@@ -34,6 +38,7 @@ class ResourcesDeleteController extends AppController
      * @throws NotFoundException If the user does not have access to the resource.
      * @throws ForbiddenException If the user does not have the permission to delete the resource.
      * @throws BadRequestException If the resource id is not a valid uuid.
+     * @throws InternalErrorException if the resource could not be saved for other reasons
      * @return void
      */
     public function delete($id)
@@ -52,10 +57,19 @@ class ResourcesDeleteController extends AppController
             throw new NotFoundException(__('The resource does not exist.'));
         }
 
-        // Save the entity
-        $this->Resources->softDelete($this->User->id(), $resource);
-        $this->_handleDeleteError($resource);
+        // Get the list of users who have access to the resource
+        // useful to do now to notify users later, since it wont be possible to after delete
+        $Users = TableRegistry::get('Users');
+        $options = ['contain' => ['Roles'], 'filter' => ['has-access' => [$resource->id]]];
+        $users = $Users->findIndex(Role::USER, $options)->all();
 
+        // Update the entity to delete=1 and drop associated permissions
+        if (!$this->Resources->softDelete($this->User->id(), $resource)) {
+            $this->_handleDeleteError($resource);
+            throw new InternalErrorException(__('Could not delete the resource. Please try again later.'));
+        }
+
+        $this->_notifyUser($resource, $users);
         $this->success(__('The resource was deleted'));
     }
 
@@ -73,17 +87,33 @@ class ResourcesDeleteController extends AppController
         if (empty($errors)) {
             return;
         }
-
         if (isset($errors['deleted']['is_not_soft_deleted'])) {
             throw new NotFoundException(__('The resource does not exist.'));
-        } elseif (isset($errors['id']['has_access'])) {
+        }
+        if (isset($errors['id']['has_access'])) {
             // If the user has a read access return a 403, otherwise return a 404 to avoid data leak.
             if ($this->Resources->hasAccess($this->User->id(), $resource->id)) {
                 throw new ForbiddenException(__('You do not have the permission to delete this resource.'));
             }
             throw new NotFoundException(__('The resource does not exist.'));
         }
-
         throw new ValidationRuleException(__('Could not delete the resource.'), $errors, $this->Resources);
+    }
+
+    /**
+     * Send email notification
+     *
+     * @param \App\Model\Entity\Resource $resource Resource
+     * @param \Cake\Datasource\ResultSetInterface $users list of User entity who had access to the resource
+     * @return void
+     */
+    protected function _notifyUser(\App\Model\Entity\Resource $resource, \Cake\Datasource\ResultSetInterface $users)
+    {
+        $event = new Event('ResourcesDeleteController.delete.success', $this, [
+            'resource' => $resource,
+            'deletedBy' => $this->User->id(),
+            'users' => $users
+        ]);
+        $this->getEventManager()->dispatch($event);
     }
 }
